[pages] feat: added page structure
